+++ src/s_sound.h
@@ -85,7 +85,8 @@ void S_Init(void);
 // Shut down sound
 void S_Shutdown(void);
 
-void S_StopSound(const sfxnum_t sfxnum);
+void S_StopSoundEffect(const sfxnum_t sfxnum);
+void S_StopSound(const mobj_t *origin);
 void S_StopSounds(void);
 
 //