+++ releasenotes.md
@@ -46,6 +46,7 @@
 * The support for *Direct 3D 11* has been removed. The `vid_scaleapi` CVAR can no longer be set to `direct3d9` or `direct3d11`, but `direct3d` instead.
 * The positioning of tall skies has improved in some instances when the `freelook` CVAR is `off`.
 * Improvements have been made to fixing the offsets of sprites when the `r_fixspriteoffsets` CVAR is `on`.
+* Playing sound effects in full as things are removed from a map can now be disabled using the new `s_fullsfx` CVAR, which is `on` by default and `off` when vanilla mode is enabled.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)